requirements.txt
boto3==1.39.4
gTTS==2.5.4
moviepy==2.2.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
pillow==11.2.1
librosa==0.11.0
soundfile==0.13.1
torchaudio==2.6.0
chatterbox-tts==0.1.2
whisper
scikit-learn==1.6.1
torch==2.6.0
transformers==4.46.3
numpy==2.0.2
scipy==1.13.1
openai==0.28.0
openai-whisper
python-dotenv==1.1.1
python-multipart==0.0.20
pydantic==2.11.7
PyYAML==6.0.2
SQLAlchemy>=2.0
psycopg2-binary>=2.9
tenacity==8.2.3
google-genai
psutil>=5.9.0
